rename and format settings files
